refactoring like there's no tomorrow - WIP/ show s3 articles on dash
