+++ bin/lib/utils/utils.js
@@ -69,9 +69,40 @@ function maybeAppendDot(text) {
 	return text + (text.endsWith('?') ? '' : '.');
 }
 
+function formatOutput(translations, hasStandfirst, sourceOnly = false) {
+	// convert \n\n-separated blocks of text into <p>-wrapped blocks of text
+	translations.translatorNames.map(translatorName => {
+		let textWithParas;
+		if (translations.texts[translatorName].hasOwnProperty('error')) {
+			textWithParas = translations.texts[translatorName];
+		} else {
+
+			if(!sourceOnly || (sourceOnly && translatorName === 'original')) {
+				const textWithBackslashNs = translations.texts[translatorName];
+				const paras = textWithBackslashNs.split('\n\n').map((para, index) => {
+					if (index === 0) {
+						return `<h1>${para}</h1>`;	
+					} 
+
+					if (index === 1 && hasStandfirst) {
+						return `<h2>${para}</h2>`;
+					}
+
+					return para.length > 0 ?`<p>${para}</p>`:'';
+				});
+				textWithParas = paras.join('\n');
+				translations.texts[translatorName] = textWithParas;
+			}
+		}
+	});
+
+	return translations;
+}
+
 module.exports = {
 	extractUser: getS3OUserFromCookie,
 	splitTextIntoChunks: checkAndSplitText,
 	pauseForMillis: pauseForMillis,
-	maybeAppendDot: maybeAppendDot
+	maybeAppendDot: maybeAppendDot,
+	formatOutput: formatOutput
 };
